+++ public/components/notebooks/components/notebook.tsx
@@ -35,6 +35,7 @@ import { useHistory, useLocation } from 'react-router-dom';
 import { Subscription, timer } from 'rxjs';
 import { switchMap, takeWhile } from 'rxjs/operators';
 
+import { ActionMetadata } from 'common/constants/actions';
 import { CoreStart, MountPoint, SavedObjectsStart } from '../../../../../../src/core/public';
 import { DashboardStart } from '../../../../../../src/plugins/dashboard/public';
 import { DataSourceManagementPluginSetup } from '../../../../../../src/plugins/data_source_management/public';
@@ -56,12 +57,15 @@ import {
 import { Paragraphs } from './paragraph_components/paragraphs';
 import { ContextPanel } from './context_panel';
 import { NotebookContextProvider } from '../context_provider/context_provider';
-import { getMLCommonsTask, executeMLCommonsAgent, getMLCommonsConfig } from '../../../utils/ml_commons_apis';
+import {
+  getMLCommonsTask,
+  executeMLCommonsAgent,
+  getMLCommonsConfig,
+} from '../../../utils/ml_commons_apis';
 import { parseParagraphOut } from '../../../utils/paragraph';
 import { isStateCompletedOrFailed } from '../../../utils/task';
 import { constructDeepResearchParagraphOut } from '../../../../common/utils/paragraph';
-import { ActionMetadata, InputPanel } from './input_panel';
-import { clog } from '../../../../../../src/plugins/expressions';
+import { InputPanel } from './input_panel';
 
 const ParagraphTypeDeepResearch = 'DEEP_RESEARCH';
 
@@ -125,10 +129,20 @@ export function Notebook({
   const taskSubscriptions = useRef(new Map<string, Subscription>());
 
   const executeActionSelectionAgent = async (input: string, actionsMetadata: ActionMetadata[]) => {
-    const {configuration: {agent_id: actionSelectionAgentId}} = await getMLCommonsConfig({http: http, configName: "action-selection-agent"})
-    const result = await executeMLCommonsAgent({ http: http, agentId: actionSelectionAgentId, dataSourceId: dataSourceMDSId ?? undefined, parameters: {actions_metadata: actionsMetadata.map((acion) => JSON.stringify(acion)).join(','), input_question: input} })
+    const {
+      configuration: { agent_id: actionSelectionAgentId },
+    } = await getMLCommonsConfig({ http, configName: 'action-selection-agent' });
+    const result = await executeMLCommonsAgent({
+      http,
+      agentId: actionSelectionAgentId,
+      dataSourceId: dataSourceMDSId ?? undefined,
+      parameters: {
+        actions_metadata: actionsMetadata.map((acion) => JSON.stringify(acion)).join(','),
+        input_question: input,
+      },
+    });
     return result;
-  }
+  };
 
   const toggleReportingLoadingModal = (show: boolean) => {
     setIsReportingLoadingModalOpen(show);
@@ -1310,10 +1324,11 @@ export function Notebook({
           </EuiPageBody>
           <EuiSpacer />
           <InputPanel
-              onCreateParagraph={handleCreateParagraph}
-              selectAction={(input: string, actionsMetadata: ActionMetadata[]) => executeActionSelectionAgent(input, actionsMetadata)
-              }
-            />
+            onCreateParagraph={handleCreateParagraph}
+            selectAction={(input: string, actionsMetadata: ActionMetadata[]) =>
+              executeActionSelectionAgent(input, actionsMetadata)
+            }
+          />
         </EuiPage>
         {isModalVisible && modalLayout}
       </>
