+++ public/utils/ml_commons_apis.ts
@@ -195,70 +195,3 @@ export const getMLCommonsConfig = ({
     signal,
     dataSourceId,
   });
-
-const actionsMetadata = [
-  {
-    id: 'PPL',
-    title: 'PPL',
-    description:
-      'display PPL block: Piped Processing Language (PPL) is a query language that focuses on processing data in a sequential, step-by-step manner.',
-    input_metadata: {
-      ppl: {
-        type: 'string',
-        description: 'ppl query',
-        required: true,
-      },
-    },
-  },
-  {
-    id: 'DEEP_RESEARCH_AGENT',
-    title: 'PlanAndExecuteAgent',
-    description:
-      'display PlanAndExecuteAgent block: PlanAndExecuteAgent is capable of breaking down complex tasks into simple steps and re-evaluating the steps based on intermediate results.',
-    input_metadata: {
-      question: {
-        type: 'string',
-        description: 'user question',
-        required: true,
-      },
-    },
-  },
-  {
-    id: 'VISUALIZATION',
-    title: 'Visualization',
-    description:
-      'display visualization block: user can select existing visualization and display it in the block.',
-  },
-  {
-    id: 'MARKDOWN',
-    title: 'Markdown',
-    description:
-      'display markdown block: If the input follows markdown syntax, use the markdown editor to create formatted text.',
-    input_metadata: {
-      markdown_text: {
-        type: 'string',
-        description: 'markdown text',
-        required: true,
-      },
-    },
-  },
-];
-
-// export const selectAction = async ({
-//   http,
-//   inputQuestion,
-// }: {
-//   http: CoreStart['http'];
-//   inputQuestion: string;
-// }) => 
-//   callApiWithProxy({
-//     http,
-//     method: 'POST',
-//     path: OPENSEARCH_ML_COMMONS_API.agentExecute.replace('{agentId}', agentId),
-//     body: JSON.stringify({
-//       parameters: {
-//         actions_metadata: actionsMetadata.map((acion) => JSON.stringify(acion)).join(','),
-//         input_question: inputQuestion,
-//       },
-//     }),
-//   });
